refactor!: Move SettingControlBindings to GitUI
from GitExtensions.Extensibility,
make ISetting.CreateControlBinding optional,
remove TextBoxValidationColors again

Refs: #13031
