+++ src/app/GitExtensions.Extensibility/Properties/AssemblyInfo.cs
@@ -2,3 +2,4 @@
 
 [assembly: InternalsVisibleTo("GitExtensions.Extensibility.Tests")]
 [assembly: InternalsVisibleTo("GitCommands")]
+[assembly: InternalsVisibleTo("GitUI")] // NumberSetting<T>.TryConvertFromString is used in GitUI's NumberSettingTextBoxBinding.
