+++ src/app/GitExtensions.Extensibility/Settings/CredentialsSetting.cs
@@ -42,65 +42,5 @@ public class CredentialsSetting : ICredentialsManager, ISetting
         _credentialsManager.SetCredentials(settings.SettingLevel, Name, newCredentials);
     }
 
-    public ISettingControlBinding CreateControlBinding()
-    {
-        return new CredentialsControlBinding(this, CustomControl);
-    }
-
     public void Save() => _credentialsManager.Save();
-
-    private class CredentialsControlBinding : SettingControlBinding<CredentialsSetting, CredentialsControl>
-    {
-        public CredentialsControlBinding(CredentialsSetting setting, CredentialsControl? control)
-            : base(setting, control)
-        {
-        }
-
-        public override CredentialsControl CreateControl()
-        {
-            Setting.CustomControl = new CredentialsControl();
-            return Setting.CustomControl;
-        }
-
-        public override void LoadSetting(SettingsSource settings, CredentialsControl control)
-        {
-            if (SettingLevelSupported(settings.SettingLevel))
-            {
-                NetworkCredential credentials = Setting.GetValueOrDefault(settings);
-                control.UserName = credentials.UserName;
-                control.Password = credentials.Password;
-                control.Enabled = true;
-            }
-            else
-            {
-                control.UserName = string.Empty;
-                control.Password = string.Empty;
-                control.Enabled = false;
-            }
-        }
-
-        public override void SaveSetting(SettingsSource settings, CredentialsControl control)
-        {
-            if (SettingLevelSupported(settings.SettingLevel))
-            {
-                Setting.SaveValue(settings, control.UserName, control.Password);
-
-                // Reload actual settings.
-                LoadSetting(settings, control);
-            }
-        }
-
-        private static bool SettingLevelSupported(SettingLevel settingLevel)
-        {
-            switch (settingLevel)
-            {
-                case SettingLevel.Global:
-                case SettingLevel.Local:
-                case SettingLevel.Effective:
-                    return true;
-            }
-
-            return false;
-        }
-    }
 }
