+++ src/app/GitUI/CommandsDialogs/SettingsDialog/Pages/DetailedSettingsPage.cs
@@ -1,6 +1,7 @@
 ﻿using GitCommands;
 using GitCommands.Settings;
 using GitExtensions.Extensibility.Settings;
+using GitUI.SettingControlBindings;
 
 namespace GitUI.CommandsDialogs.SettingsDialog.Pages;
 
@@ -13,9 +14,9 @@ public partial class DetailedSettingsPage : DistributedSettingsPage
         InitializeComponent();
         InitializeComplete();
 
-        _controlBindings = [DetailedSettings.GetRemoteBranchesDirectlyFromRemote.CreateControlBinding(chkRemotesFromServer),
-                            DetailedSettings.AddMergeLogMessages.CreateControlBinding(addLogMessages),
-                            DetailedSettings.MergeLogMessagesCount.CreateControlBinding(nbMessages)];
+        _controlBindings = [SettingControlBindingsProvider.CreateControlBinding(DetailedSettings.GetRemoteBranchesDirectlyFromRemote, chkRemotesFromServer),
+                            SettingControlBindingsProvider.CreateControlBinding(DetailedSettings.AddMergeLogMessages, addLogMessages),
+                            SettingControlBindingsProvider.CreateControlBinding(DetailedSettings.MergeLogMessagesCount, nbMessages)];
     }
 
     public static SettingsPageReference GetPageReference()
