+++ src/app/GitUI/SettingControlBindings/NumberSettingControlBinding.cs
@@ -1,0 +1,10 @@
+﻿namespace GitUI.SettingControlBindings;
+
+internal static class NumberSettingControlBinding
+{
+    /// <summary>
+    ///  The text to be shown as a hint in the input control when the value
+    ///  of a <see cref="GitExtensions.Extensibility.Settings.NumberSetting{T}"/> is null.
+    /// </summary>
+    public static string PlaceholderText { get; set; } = "";
+}
