+++ tests/app/UnitTests/GitUI.Tests/SettingControlBindings/StringSettingControlBindingTests.cs
@@ -1,10 +1,11 @@
 ﻿using GitExtensions.Extensibility.Settings;
+using GitUI.SettingControlBindings;
 using NSubstitute;
 
-namespace GitUIPluginInterfacesTests.Settings;
+namespace GitUITests.SettingControlBindings;
 
 [TestFixture]
-public sealed class StringSettingTests
+public sealed class StringSettingControlBindingTests
 {
     private const string SettingName = "TestSetting";
     private const string DefaultValue = "default";
@@ -14,7 +15,8 @@ public sealed class StringSettingTests
     {
         StringSetting setting = new(SettingName, DefaultValue);
         using TextBox textBox = new();
-        ISettingControlBinding binding = setting.CreateControlBinding();
+        setting.CustomControl = textBox;
+        ISettingControlBinding binding = SettingControlBindingsProvider.CreateControlBinding(setting);
         SettingsSource settingsSource = Substitute.For<SettingsSource>();
 
         textBox.Text = string.Empty;
@@ -29,7 +31,7 @@ public sealed class StringSettingTests
         StringSetting setting = new(SettingName, DefaultValue);
         using TextBox textBox = new();
         setting.CustomControl = textBox;
-        ISettingControlBinding binding = setting.CreateControlBinding();
+        ISettingControlBinding binding = SettingControlBindingsProvider.CreateControlBinding(setting);
         SettingsSource settingsSource = Substitute.For<SettingsSource>();
         settingsSource.SettingLevel.Returns(SettingLevel.Effective);
         settingsSource.GetValue(SettingName).Returns((string?)null);
@@ -46,7 +48,7 @@ public sealed class StringSettingTests
         StringSetting setting = new(SettingName, DefaultValue);
         using TextBox textBox = new();
         setting.CustomControl = textBox;
-        ISettingControlBinding binding = setting.CreateControlBinding();
+        ISettingControlBinding binding = SettingControlBindingsProvider.CreateControlBinding(setting);
         SettingsSource settingsSource = Substitute.For<SettingsSource>();
         settingsSource.SettingLevel.Returns(settingLevel);
         settingsSource.GetValue(SettingName).Returns((string?)null);
@@ -62,7 +64,7 @@ public sealed class StringSettingTests
         StringSetting setting = new(SettingName, DefaultValue);
         using TextBox textBox = new();
         setting.CustomControl = textBox;
-        ISettingControlBinding binding = setting.CreateControlBinding();
+        ISettingControlBinding binding = SettingControlBindingsProvider.CreateControlBinding(setting);
         SettingsSource settingsSource = Substitute.For<SettingsSource>();
         settingsSource.SettingLevel.Returns(SettingLevel.Effective);
         settingsSource.GetValue(SettingName).Returns((string?)null);
@@ -80,7 +82,7 @@ public sealed class StringSettingTests
         StringSetting setting = new(SettingName, DefaultValue);
         using TextBox textBox = new();
         setting.CustomControl = textBox;
-        ISettingControlBinding binding = setting.CreateControlBinding();
+        ISettingControlBinding binding = SettingControlBindingsProvider.CreateControlBinding(setting);
         SettingsSource settingsSource = Substitute.For<SettingsSource>();
 
         textBox.Text = value;
@@ -96,7 +98,7 @@ public sealed class StringSettingTests
         StringSetting setting = new(SettingName, DefaultValue);
         using TextBox textBox = new();
         setting.CustomControl = textBox;
-        ISettingControlBinding binding = setting.CreateControlBinding();
+        ISettingControlBinding binding = SettingControlBindingsProvider.CreateControlBinding(setting);
         SettingsSource settingsSource = Substitute.For<SettingsSource>();
         settingsSource.SettingLevel.Returns(settingLevel);
